Fix clan types (#3)
Fix clan types (#3)

What changed?
===================

This commit fixes return types for the clan and clanlist routes as well as the home route. It also fixes the error that occurs when there are no members in the member list.

Why?
===================

To make sure the functions return the appropriate value by checking against the type.